traceability-controller/requirements.txt
annotated-types==0.6.0
anoncreds==0.2.0.dev6
anyio==3.7.1
aries-askar==0.3.0
async-timeout==4.0.3
base58==2.1.1
bitarray==2.9.1
bitstring==4.1.4
black==23.12.1
cached-property==1.5.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
exceptiongroup==1.2.0
fastapi==0.105.0
frozendict==2.3.10
h11==0.14.0
idna==3.6
inflection==0.5.1
lxml==5.0.0
marshmallow==3.20.2
morphys==1.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
py-multibase==1.0.3
py-multicodec==0.2.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
PyLD==2.0.3
python-baseconv==1.2.2
python-dotenv==1.0.0
python-multipart==0.0.6
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uuid==1.30
uvicorn==0.25.0
validators==0.22.0
varint==1.0.2
